requirements.txt
Django==4.2
python-dotenv
djangorestframework
psycopg2-binary
ca-vntl-helper==1.1.6
wikipedia
openai
langchain
langchain_openai
langchain_google_genai
langchain_community
langchain-postgres
django-rest-swagger
drf-yasg
python-dotenv
setuptools
pgvector
duckduckgo-search
django-cors-headers
djangorestframework_simplejwt
pymupdf
pytesseract
pdfminer.six
Pillow
neo4j
neomodel
langchain_experimental
grpcio
grpcio-tools
djangogrpcframework
grpc-django